requirements.txt
django-toolbelt==0.0.1
static3==0.5.1
wsgiref==0.1.2
python-social-auth==0.2.1
pygments==2.0.1